requirements.txt
asgiref==3.5.2
attrs==22.1.0
beautifulsoup4==4.11.1
boto3==1.24.65
botocore==1.27.65
cached-property==1.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.4
django-admin-thumbnails==0.2.6
django-appconf==1.0.5
django-ckeditor==6.4.1
django-compressor==4.0
django-filter==22.1
django-htmlmin==0.11.0
django-js-asset==2.0.0
django-storages==1.13.1
django-widget-tweaks==1.4.12
html5lib==1.1
idna==3.3
ippanel==1.1.0
isodate==0.6.1
jmespath==1.0.1
lxml==4.9.1
Pillow==9.1.1
platformdirs==2.5.2
psycopg2==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2
rcssmin==1.1.0
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
rjsmin==1.2.0
s3transfer==0.6.0
six==1.16.0
sorl-thumbnail==12.8.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
webencodings==0.5.1
zeep==4.1.0
